+++ CHANGELOG.md
@@ -7,6 +7,8 @@ New features:
 - When you click any non-conversation file in the mindspace tree it now opens it in the dynamic wiki view (just as drag
   and drop does).  Conversations open in a conversation tab.  If you double-click a file it opens in an editor.
 - All "hidden" files except for ".humbug" are now shown in the mindspace tree.
+- There are new icons on message boxes.
+- The conversation input box now has a submit icon.
 
 Bug fixes: